requirements.txt
# 🕵️ Desktop OSINT Investigation Suite - Dependencies
# LakyLuk Enhanced Edition - 27.9.2025

# Core Framework
aiohttp>=3.8.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
sqlalchemy>=2.0.0

# GUI Framework
tkinter-tooltip>=1.2.1

# Web Automation & Scraping
selenium>=4.15.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
fake-useragent>=1.4.0

# AI Integration
anthropic>=0.7.0
openai>=1.3.0
google-generativeai>=0.3.0

# Data Processing & Analysis
pandas>=2.1.0
numpy>=1.24.0
networkx>=3.1.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
plotly>=5.17.0
seaborn>=0.12.0

# Security & Encryption
cryptography>=41.0.0
requests[socks]>=2.31.0

# Reporting & Export
reportlab>=4.0.0
openpyxl>=3.1.0
jinja2>=3.1.0
fpdf2>=2.7.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Network & Proxy
aiohttp-socks>=0.8.0
python-socks>=2.3.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
flake8>=6.0.0

# Optional: Local LLM Support
# ollama>=0.1.0  # Uncomment if using local LLM

# Optional: Additional OSINT Tools
# shodan>=1.28.0  # Uncomment if using Shodan API
# censys>=2.2.0   # Uncomment if using Censys API